requirements.txt
# Django e Framework Web
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Dash e Visualização
dash==2.14.2
plotly==5.17.0
dash-bootstrap-components==1.5.0

# Processamento de Dados
pandas==2.1.4
numpy==1.24.4
yfinance==0.2.28

# Banco de Dados
psycopg2-binary==2.9.9
SQLAlchemy==1.4.53

# Orquestração
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# Segurança e Autenticação
PyJWT==2.8.0
cryptography==39.0.2
python-dotenv==1.0.0

# Logging e Observabilidade
structlog==23.2.0
python-json-logger==2.0.7

# Utilitários
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Desenvolvimento e Testes
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Docker e Infraestrutura
gunicorn==21.2.0 